feat(dashboard): Refactored Dashboard Route
- Moved Dashboard buttons into a separate component to improve readability

- Created new component in Features/Dashboard/SendModal
- Uses previous modal logic and state, but moved into SendModal

- Added SelectGradeOptions component to Features/Utils
- Component allows us to re-use option component for v1 through v17

Fixes #201
