+++ frontend/src/Features/Utils/SelectGradeOptions.jsx
@@ -1,0 +1,11 @@
+export default function SelectGradeOptions(){
+
+    
+        const versionOptions = [];
+        for (let i = 1; i <= 17; i++) {
+            versionOptions.push(<option key={`v${i}`} value={`v${i}`}>v{i}</option>);
+        }
+   
+        return versionOptions
+   
+}
